[justcode 이상윤] Layered Pattern 전체 적용 - 각각의 router, controller, service, models 는 유저, 게시글, 코멘트의 기능들로 분류할 수 있다. - 따라서 해당 기능들에 따라 파일들을 분류하였음.
