+++ sangyun/3rd_week/src/models/post.dao.ts
@@ -1,115 +1,4 @@
-import { DataSource } from 'typeorm';
-
-const dataSource = new DataSource({
-  type: process.env.TYPEORM_CONNECTION,
-  host: process.env.TYPEORM_HOST,
-  port: Number(process.env.TYPEORM_PORT),
-  username: process.env.TYPEORM_USERNAME,
-  password: process.env.TYPEORM_PASSWORD,
-  database: process.env.TYPEORM_DATABASE,
-});
-
-dataSource.initialize().then(() => {
-  console.log("Data Source has been initialized!");
-});
-
-
-async function findUserByEmail(email : string) : Promise<UserInfo> {
-  const [userInfo] = await dataSource.query(`
-    SELECT
-      id,
-      password
-    FROM users
-    WHERE
-    email = ?
-  `, [email]);
-  return userInfo as UserInfo;
-}
-
-async function findUserById(id : string) : Promise<UserInfo> {
-  const [userInfo] = await dataSource.query(`
-    SELECT
-      id,
-      password,
-      email,
-      nickname
-    FROM users
-    WHERE
-    id = ?
-  `, [id]);
-  return userInfo as UserInfo;
-}
-
-async function getAllUser() {
-  const allUsers = await dataSource.query(`
-    SELECT
-      id,
-      email,
-      nickname
-    FROM users
-  `);
-  return allUsers;
-}
-
-async function addUser(userInfo : {email: string, password: string, nickname: string, profile_image: string}) {
-  const {email, nickname, password, profile_image} = userInfo;
-  return await dataSource.query(
-    `INSERT INTO users(
-                        email,
-                        nickname,
-                        password,
-                        profile_image
-                      ) VALUES (?, ?, ?, ?);
-                      `,
-    [email, nickname, password, profile_image]
-  )
-}
-
-async function addCommentOnPost(userId: string, postId: string, comment: string) {
-  return await dataSource.query(`
-    INSERT INTO comments(
-      comment,
-      posting_id,
-      user_id
-      )
-    VALUES(?, ?, ?);`
-  , [comment, postId, userId])
-}
-
-async function findCommentById(commentId: string) {
-  const [comment] = await dataSource.query(`
-      SELECT
-        comments.comment,
-        postings.id AS post_id,
-        comments.created_at,
-        comments.updated_at,
-        users.nickname as user_nickname,
-        users.id as user_id
-      FROM comments
-      JOIN users    ON comments.user_id     = users.id
-      JOIN postings ON comments.posting_id  = postings.id
-      WHERE comments.id = ?`
-      , [commentId]
-    );
-  return comment as {comment: string, poast_id: string, created_at: string, updated_at: string, user_nickname: string, user_id: string, };
-}
-
-async function updateComment(commentId: string, comment: string) {
-  await dataSource.query(`
-      UPDATE comments
-        SET comment = ?
-      WHERE id = ?`
-      , [comment, commentId]
-    )
-}
-
-async function deleteComment(commentId: string) {
-  await dataSource.query(`
-    DELETE FROM
-      comments
-    WHERE id = ?`
-    ,[commentId])
-}
+import dataSource from './database';
 
 async function addPost(contents: string, image_url: string, userId: string) {
   await dataSource.transaction(async (transactionalEntityManager) => {
@@ -323,14 +212,6 @@ async function addLikePost(userId: string, postId: string) {
 }
 
 export default {
-  findUserByEmail,
-  findUserById,
-  getAllUser,
-  addUser,
-  addCommentOnPost,
-  findCommentById,
-  updateComment,
-  deleteComment,
   addPost,
   getAllPost,
   getPostByPostId,
