+++ sangyun/3rd_week/src/routers/index.ts
@@ -1,7 +1,13 @@
 import express from 'express';
-import allRouter from './all-router';
+import cmtRouter from './comment.router';
+import userRouter from './user.router';
+import postRouter from './post.router';
+import middleware from '../middlewares/middleware';
 
 const router = express.Router();
-router.use('', allRouter);
+router.use('', cmtRouter);
+router.use('', userRouter);
+router.use('', postRouter);
+router.use(middleware.errorHandler);
 
 export default router;
