+++ sangyun/3rd_week/src/services/index.ts
@@ -1,0 +1,11 @@
+import cmtSvc from './comment.service';
+import userSvc from './user.service';
+import postSvc from './post.service';
+
+const svc_set = {
+  cmtSvc,
+  postSvc,
+  userSvc,
+}
+
+export default svc_set;
